+++ packages/core/include/combo/menu.h
@@ -8,9 +8,11 @@
 #define MENU_SOULS_OOT_ENEMY    2
 #define MENU_SOULS_OOT_BOSS     3
 #define MENU_SOULS_OOT_NPC      4
-#define MENU_SOULS_MM_ENEMY     5
-#define MENU_SOULS_MM_BOSS      6
-#define MENU_MAX                7
+#define MENU_SOULS_OOT_MISC     5
+#define MENU_SOULS_MM_ENEMY     6
+#define MENU_SOULS_MM_BOSS      7
+#define MENU_SOULS_MM_MISC      8
+#define MENU_MAX                9
 
 typedef struct GameState_Play GameState_Play;