+++ packages/core/lib/combo/logic/world-transform.ts
@@ -202,11 +202,14 @@ const ITEM_POOL_PLENTIFUL = new Set([
   Items.SHARED_WALLET,
   Items.SHARED_SKELETON_KEY,
   ...ItemGroups.OOT_SOULS_ENEMY,
-  ...ItemGroups.MM_SOULS_ENEMY,
-  ...ItemGroups.OOT_SOULS_BOSS,
-  ...ItemGroups.MM_SOULS_BOSS,
   ...ItemGroups.OOT_SOULS_NPC,
+  ...ItemGroups.OOT_SOULS_BOSS,
+  ...ItemGroups.OOT_SOULS_MISC,
+  ...ItemGroups.MM_SOULS_ENEMY,
+  ...ItemGroups.MM_SOULS_BOSS,
+  ...ItemGroups.MM_SOULS_MISC,
   ...ItemGroups.SHARED_SOULS_ENEMY,
+  ...ItemGroups.SHARED_SOULS_MISC,
   Items.OOT_BUTTON_A,
   Items.OOT_BUTTON_C_RIGHT,
   Items.OOT_BUTTON_C_LEFT,
@@ -849,6 +852,10 @@ export class LogicPassWorldTransform {
       this.shareItems(SharedItemGroups.SOULS_ENEMY, 'max');
     }
 
+    if (settings.sharedSoulsMisc) {
+      this.shareItems(SharedItemGroups.SOULS_MISC, 'max');
+    }
+
     if (settings.sharedSkeletonKey) {
       this.shareItems(SharedItemGroups.SKELETON_KEY, 'max');
     }
@@ -1082,10 +1089,12 @@ export class LogicPassWorldTransform {
 
     /* Add souls */
     if (settings.soulsEnemyOot) this.addItems(ItemGroups.OOT_SOULS_ENEMY);
-    if (settings.soulsEnemyMm) this.addItems(ItemGroups.MM_SOULS_ENEMY);
     if (settings.soulsBossOot) this.addItems(ItemGroups.OOT_SOULS_BOSS);
-    if (settings.soulsBossMm) this.addItems(ItemGroups.MM_SOULS_BOSS);
     if (settings.soulsNpcOot) this.addItems(ItemGroups.OOT_SOULS_NPC);
+    if (settings.soulsMiscOot) this.addItems(ItemGroups.OOT_SOULS_MISC);
+    if (settings.soulsEnemyMm) this.addItems(ItemGroups.MM_SOULS_ENEMY);
+    if (settings.soulsBossMm) this.addItems(ItemGroups.MM_SOULS_BOSS);
+    if (settings.soulsMiscMm) this.addItems(ItemGroups.MM_SOULS_MISC);
 
     /* Add skeleton keys */
     if (settings.skeletonKeyOot) {