+++ packages/core/src/mm/actors.c
@@ -18,6 +18,11 @@ void EnTab_AfterTextBox(Actor* this, GameState_Play* play, s16 messageId);
 
 static Actor* sByteCodeActor;
 
+static int opt(int x)
+{
+    return x > 0 ? 1 : -1;
+}
+
 /* TODO: Patch every bytecode and remove this */
 static s16 convertGi(s16 initial)
 {
@@ -267,8 +272,10 @@ static int canSpawnSoul(GameState_Play* play, s16 actorId, u16 variable)
         return comboHasSoulMm(GI_MM_SOUL_ENEMY_REDEAD_GIBDO);
     case AC_EN_SW:
         if (variable & 0x03)
-            return 1;
+            return opt(comboHasSoulMm(GI_MM_SOUL_MISC_GS));
         return comboHasSoulMm(GI_MM_SOUL_ENEMY_SKULLWALLTULA);
+    case AC_OBJ_MAKEKINSUTA:
+        return opt(comboHasSoulMm(GI_MM_SOUL_MISC_GS));
     case AC_EN_SB:
         return comboHasSoulMm(GI_MM_SOUL_ENEMY_SHELL_BLADE);
     case AC_EN_RR:
@@ -353,6 +360,10 @@ static int canSpawnSoul(GameState_Play* play, s16 actorId, u16 variable)
     case AC_EN_HINT_SKB:
     case AC_EN_RAIL_SKB:
         return comboHasSoulMm(GI_MM_SOUL_ENEMY_STALCHILD);
+    case AC_EN_SELLNUTS:
+    case AC_EN_SCOPENUTS:
+    case AC_EN_AKINDONUTS:
+        return opt(comboHasSoulMm(GI_MM_SOUL_MISC_BUSINESS_SCRUB));
     default:
         return 1;
     }